move compile rules to separate file
[ci skip-rust]

[ci skip-build-wheels]
