+++ src/python/pants/bsp/rules.py
@@ -2,7 +2,6 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 from __future__ import annotations
 
-import time
 from dataclasses import dataclass
 
 from pants.base.build_root import BuildRoot
@@ -12,11 +11,7 @@ from pants.bsp.spec import (
     BuildServerCapabilities,
     BuildTarget,
     BuildTargetIdentifier,
-    CompileParams,
     CompileProvider,
-    CompileReport,
-    CompileResult,
-    CompileTask,
     DebugProvider,
     InitializeBuildParams,
     InitializeBuildResult,
@@ -26,15 +21,12 @@ from pants.bsp.spec import (
     SourcesItem,
     SourcesParams,
     SourcesResult,
-    StatusCode,
-    TaskFinishParams,
-    TaskId,
-    TaskStartParams,
     TestProvider,
     WorkspaceBuildTargetsParams,
     WorkspaceBuildTargetsResult,
 )
 from pants.bsp.types import BSPLanguageSupport
+from pants.bsp.util_rules.compile import rules as bsp_compile_rules
 from pants.build_graph.address import AddressInput
 from pants.engine.internals.selectors import Get, MultiGet
 from pants.engine.internals.session import SessionValues
@@ -211,52 +203,11 @@ async def bsp_build_target_sources(request: SourcesParams) -> SourcesResult:
     return SourcesResult(items=tuple(si.sources_item for si in sources_items))
 
 
-# -----------------------------------------------------------------------------------------------
-# Compile Request
-# See https://build-server-protocol.github.io/docs/specification.html#compile-request
-# -----------------------------------------------------------------------------------------------
-
-
-class CompileRequestHandlerMapping(BSPHandlerMapping):
-    method_name = "buildTarget/compile"
-    request_type = CompileParams
-    response_type = CompileResult
-
-
-@rule
-async def bsp_compile_request(request: CompileParams, bsp_context: BSPContext) -> CompileResult:
-    origin_id = request.origin_id or "compile-task"
-    for i, target in enumerate(request.targets):
-        task_id = TaskId(id=f"{origin_id}-{i}")
-        bsp_context.notify_client(
-            TaskStartParams(
-                task_id=task_id,
-                event_time=int(time.time() * 1000),
-                data=CompileTask(target=target),
-            )
-        )
-
-    for i, target in enumerate(request.targets):
-        task_id = TaskId(id=f"{origin_id}-{i}")
-        bsp_context.notify_client(
-            TaskFinishParams(
-                task_id=task_id,
-                status=StatusCode.ERROR,
-                data=CompileReport(target=target, origin_id=origin_id),
-            )
-        )
-
-    return CompileResult(
-        origin_id=request.origin_id,
-        status_code=1,
-    )
-
-
 def rules():
     return (
         *collect_rules(),
+        *bsp_compile_rules(),
         UnionRule(BSPHandlerMapping, InitializeBuildHandlerMapping),
         UnionRule(BSPHandlerMapping, WorkspaceBuildTargetsHandlerMapping),
         UnionRule(BSPHandlerMapping, BuildTargetSourcesHandlerMapping),
-        UnionRule(BSPHandlerMapping, CompileRequestHandlerMapping),
     )
